# Improve chunks reader performance