```
goos: darwin
goarch: arm64
pkg: github.com/ovh/symmecrypt
               │   old.txt   │               new.txt               │
               │   sec/op    │   sec/op     vs base                │
ChunksWriter-8   38.86µ ± 1%   39.09µ ± 1%   +0.59% (p=0.001 n=10)
ChunksReader-8   4.530µ ± 2%   2.921µ ± 0%  -35.50% (p=0.000 n=10)
geomean          13.27µ        10.69µ       -19.45%

               │    old.txt    │               new.txt                │
               │     B/op      │     B/op      vs base                │
ChunksWriter-8    66.69Ki ± 0%   66.29Ki ± 0%   -0.61% (p=0.000 n=10)
ChunksReader-8   17.820Ki ± 0%   4.714Ki ± 0%  -73.55% (p=0.000 n=10)
geomean           34.47Ki        17.68Ki       -48.72%

               │  old.txt   │              new.txt               │
               │ allocs/op  │ allocs/op   vs base                │
ChunksWriter-8   464.0 ± 0%   456.0 ± 0%   -1.72% (p=0.000 n=10)
ChunksReader-8   74.00 ± 0%   44.00 ± 0%  -40.54% (p=0.000 n=10)
geomean          185.3        141.6       -23.56%
```